Made everything usize. Event system cleanup. Easy way to get components.

+++ src/crates/ferrumc_macros/src/packet.rs
@@ -161,7 +161,7 @@ pub fn bake(input: TokenStream) -> TokenStream {
     let match_arms = match_arms.into_iter();
 
     let output = quote! {
-        pub async fn handle_packet(packet_id: u8, conn_id: u32, conn_state: &crate::net::State, cursor: &mut std::io::Cursor<Vec<u8>>, state: crate::state::GlobalState) -> crate::utils::prelude::Result<()> {
+        pub async fn handle_packet(packet_id: u8, conn_id: usize, conn_state: &crate::net::State, cursor: &mut std::io::Cursor<Vec<u8>>, state: crate::state::GlobalState) -> crate::utils::prelude::Result<()> {
             match (packet_id, conn_state.as_str()) {
                 #(#match_arms)*
                 _ => tracing::warn!("No packet found for ID: 0x{:02X} in state: {}", packet_id, conn_state.as_str()),
